+++ libs/game/ui/src/lib/controls/game-controls.component.spec.ts
@@ -1,8 +1,14 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { GameControlsComponent } from './game-controls.component';
+<<<<<<< HEAD:src/app/components/controls/game-controls.component.spec.ts
+import { GameEngineService, ACHIEVEMENT_SERVICE } from '@hunt-the-bishomalo/core/services';
+import { getTranslocoTestingModule } from '@hunt-the-bishomalo/core/utils';
+import { provideRouter, Router } from '@angular/router';
+=======
 import { GameEngineService, ACHIEVEMENT_SERVICE } from '@hunt-the-bishomalo/game/data-access';
 import { getTranslocoTestingModule } from '@hunt-the-bishomalo/core';
 import { provideRouter } from '@angular/router';
+>>>>>>> df75268 (chore: temporary commit to allow merge):libs/game/ui/src/lib/controls/game-controls.component.spec.ts
 import { RouteTypes } from '@hunt-the-bishomalo/data';
 
 const mockGameService = {
@@ -22,6 +28,7 @@ const achievementServiceMock = {
 describe('GameControlsComponent', () => {
   let component: GameControlsComponent;
   let fixture: ComponentFixture<GameControlsComponent>;
+  let router: Router;
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
@@ -38,6 +45,7 @@ describe('GameControlsComponent', () => {
 
     fixture = TestBed.createComponent(GameControlsComponent);
     component = fixture.componentInstance;
+    router = TestBed.inject(Router);
     fixture.detectChanges();
   });
 
@@ -52,5 +60,46 @@ describe('GameControlsComponent', () => {
   it('should reset same level', () => {
     component.resetGame();
     expect(mockGameService.initGame).toHaveBeenCalled();
+  });
+
+  it('should call newGame and navigate to settings', () => {
+    const navigateSpy = jest.spyOn(router, 'navigate');
+    component.newGame();
+    expect(mockGameService.newGame).toHaveBeenCalled();
+    expect(navigateSpy).toHaveBeenCalledWith([RouteTypes.SETTINGS]);
+  });
+
+  it('should navigate to rules', () => {
+    const navigateSpy = jest.spyOn(router, 'navigate');
+    component.navigateToControls();
+    expect(navigateSpy).toHaveBeenCalledWith([RouteTypes.RULES]);
+  });
+
+  it('should call moveForward', () => {
+    component.moveForward();
+    expect(mockGameService.moveForward).toHaveBeenCalled();
+  });
+
+  it('should call turnLeft', () => {
+    component.turnLeft();
+    expect(mockGameService.turnLeft).toHaveBeenCalled();
+  });
+
+  it('should call turnRight', () => {
+    component.turnRight();
+    expect(mockGameService.turnRight).toHaveBeenCalled();
+  });
+
+  it('should call shootArrow', () => {
+    component.shootArrow();
+    expect(mockGameService.shootArrow).toHaveBeenCalled();
+  });
+
+  it('should toggle isVisible', () => {
+    expect(component.isVisible()).toBe(false);
+    component.toggle();
+    expect(component.isVisible()).toBe(true);
+    component.toggle();
+    expect(component.isVisible()).toBe(false);
   });
 });